+++ lib/ab_admin/menu/item.rb
@@ -1,8 +1,6 @@
 module AbAdmin
   module Menu
     class Item
-      include ::AbAdmin::Utils::EvalHelpers
-
       def initialize(title, url, options)
         @title = title.is_a?(Symbol) ? I18n.t(title, scope: [:admin, :navigation]) : title
         @url = url
@@ -10,10 +8,9 @@ module AbAdmin
       end
 
       def render(template)
-        return if @options[:if] && !call_method_or_proc_on(template, @options[:if])
-        return if @options[:unless] && call_method_or_proc_on(template, @options[:unless])
+        return unless template.option_conditions_met?(@options)
 
-        item_url = @url.is_a?(String) ? @url : call_method_or_proc_on(template, @url)
+        item_url = @url.is_a?(String) ? @url : template.instance_exec(&@url)
         active = template.request.path.split('/')[2] == item_url.split('/')[2]
 
         <<-HTML.html_safe
@@ -24,9 +21,8 @@ module AbAdmin
       private
 
       def title(template)
-        ActiveSupport::Deprecation.warn('Menu item :badge_counter option is deprecated, use :badge instead') if @options[:badge_counter]
         return @title unless @options[:badge]
-        badge = call_method_or_proc_on(template, @options[:badge])
+        badge = @options[:badge].is_a?(Symbol) ? template.public_send(@options[:badge]) : template.instance_exec(&@options[:badge])
         return @title if !badge || badge == 0
         "#{@title}&nbsp;<span class='badge badge-#{@options[:badge_type] || 'important'}'>#{badge}</span>".html_safe
       end